Develop

+++ src/main/java/com/catalogo/dragon_ball/repository/RolRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolRepository extends JpaRepository<Roles, Integer> {
+public interface RolRepository extends JpaRepository<Roles, Long> {
     Optional<Roles> findById(Long id);
 }
