+++ src/main/java/com/catalogo/dragon_ball/repository/UsuarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Users, Integer> {
+public interface UsuarioRepository extends JpaRepository<Users, Long> {
     Optional<Users> findByEmail(String email);
 }
